requirements.txt
llama-cpp-python>=0.2.90
huggingface-hub>=0.24.0
requests>=2.32.0
beautifulsoup4>=4.12.0
playwright>=1.46.0
cryptography>=43.0.0
